% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteopedia.R
\name{Calculate_AUC}
\alias{Calculate_AUC}
\title{Calculate Area Under Curve (AUC)}
\usage{
Calculate_AUC(InputDirectory, FileType = "mzML")
}
\description{
Calculates area under curve (AUC) of .mzML of .raw files from Sciex or ThermoFisher mass spectrometers
}
