% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteopedia.R
\name{Calculate_LFQ}
\alias{Calculate_LFQ}
\title{Calculate Label-Free Quantitation (LFQ)}
\usage{
Calculate_LFQ(InputData, LFQ_Col)
}
\description{
Calculates the label-free quantitation (LFQ) of proteins using the fastMaxLFQ algorithm from the iq package
}
