% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteopedia.R
\name{Check_AggregateRows}
\alias{Check_AggregateRows}
\title{Check Aggregate Rows}
\usage{
Check_AggregateRows(InputData, IDColumns)
}
\arguments{
\item{InputData}{data.table to be grouped}

\item{IDColumns}{Grouping variables of data.table}
}
\description{
Identify aggregated data.table rows using ID columns
}
