% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChromosomeLengths}
\alias{ChromosomeLengths}
\title{Data for the Standardised Analysis of Proteomic Data (Human)}
\format{
A data table with 25 rows and 2 columns.
describe{
item{Chromosome}{Chromosome ID}
item{Length}{Length of Chromosome in base-pairs}
}
}
\source{
Created by Jon Meyrick, by extracting lengths from the Human Reference Genome GENCODE V49 (GRCh38.p14)
}
\usage{
ChromosomeLengths
}
\description{
Chromosome Lengths
}
\details{
ChromosomeLengths is a data.table with the base-pair lengths of each human Chromosome
}
\examples{
data(ChromosomeLengths)
}
\keyword{datasets}
