% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Complexed_Proteins}
\alias{Complexed_Proteins}
\title{Complexed Proteins}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5961 rows and 5 columns.
}
\source{
Created by Jon Meyrick
}
\usage{
Complexed_Proteins
}
\description{
ComplexedProteins contains the names of protein complexes compiled from both Complex Portal and CORUM
}
\examples{
data(Complexed_Proteins)
}
\keyword{datasets}
