% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ComplexPortal_Complexes}
\alias{ComplexPortal_Complexes}
\title{Complex Portal Complexes}
\format{
An object of class \code{data.frame} with 8985 rows and 6 columns.
}
\source{
Created by Jon Meyrick
}
\usage{
data(ComplexPortal_Complexes)
}
\description{
Proteins assigned to complexes within the Complex Portal database
}
\examples{
data(ComplexPortal_Complexes)
}
\keyword{datasets}
