% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORUM_Complexes}
\alias{CORUM_Complexes}
\title{CORUM Complexes}
\format{
An object of class \code{data.frame} with 24560 rows and 6 columns.
}
\source{
Created by Jon Meyrick
}
\usage{
CORUM_Complexes
}
\description{
Proteins assigned to complexes within the CORUM database
}
\examples{
data(CORUM_Complexes)
}
\keyword{datasets}
