% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DegradationProfiles}
\alias{DegradationProfiles}
\title{Degradation Profiles}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3113 rows and 3 columns.
}
\source{
Created by Jon Meyrick
}
\usage{
DegradationProfiles
}
\description{
Degradation Profiles
}
\examples{
data(DegradationProfiles)
}
\keyword{datasets}
