% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteopedia.R
\name{End_Timer}
\alias{End_Timer}
\title{End Timer}
\usage{
End_Timer(Start)
}
\arguments{
\item{Start}{Start of timer called by assigning Sys.time to a variable name}
}
\value{
Message of time difference between Start and calling of funtion
}
\description{
Report time difference since defined start
}
