% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteopedia.R
\name{Generate_Metadata}
\alias{Generate_Metadata}
\title{Generate Metadata}
\usage{
Generate_Metadata(
  Condition = 0,
  Conc = 0,
  Time = 0,
  Reps = 3,
  OrderedBy = c("Condition", "Conc", "Time", "Temp", "Rep"),
  Output_Directory = "~/Desktop"
)
}
\description{
Generates metadata table using condition factor, with respect to concentration, time, and temperature variables.
Concentration, time, and temperature should be defined as Molar (M), hours (h), and degrees celsius (C), respectively
}
