% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenomicMapping}
\alias{GenomicMapping}
\title{Genomic Mapping}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 19429 rows and 5 columns.
}
\source{
Created by Jon Meyrick
}
\usage{
data(GenomicMapping)
}
\description{
Genomic Mapping
}
\examples{
data(GenomicMapping)
}
\keyword{datasets}
