% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteopedia.R
\name{\%!in\%}
\alias{\%!in\%}
\title{\%!in\%}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{Search variable}

\item{y}{String variables}
}
\value{
x variables not matching any y variables
}
\description{
Extract variables not in string
}
\examples{
c("a", "b", "c") \%!in\% c("a", "b");
"c"
}
