% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NiceColourPalette}
\alias{NiceColourPalette}
\title{NiceColourPalette is a string of hexademical colour codes which allow for plotting of 24 distinct
variable groups using clear, distinct colours}
\format{
A character string of 24 length.
}
\source{
Created by Jon Meyrick
}
\usage{
NiceColourPalette
}
\description{
NiceColourPalette is a string of hexademical colour codes which allow for plotting of 24 distinct
variable groups using clear, distinct colours
}
\keyword{datasets}
