% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnmappedProteins}
\alias{UnmappedProteins}
\title{UnmappedProteins}
\format{
A data table with 114 rows and 9 columns.
}
\source{
Created by Jon Meyrick
}
\usage{
UnmappedProteins
}
\description{
Contains proteins which cannot be mapped to Chromosomes using GenomicMapping. In most cases, this is because the Gene identifer has been changed.
}
\examples{
data(UnmappedProteins)
}
\keyword{datasets}
