% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteopedia.R
\name{Update_Proteopedia}
\alias{Update_Proteopedia}
\title{Update Proteopedia (Developer Only)}
\usage{
Update_Proteopedia(LocalDirectory)
}
\description{
For developer use, updates Proteopedia from local directory
}
