% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ThermalPalette}
\alias{ThermalPalette}
\title{ThermalPalette is a string of hexademical colour codes which allow for plotting of 46 continuous
variable groups with the colours White, Yellow, Orange, Red, Black}
\format{
A character string of 46 length.
}
\source{
Created by Jon Meyrick
}
\usage{
ThermalPalette
}
\description{
ThermalPalette is a string of hexademical colour codes which allow for plotting of 46 continuous
variable groups with the colours White, Yellow, Orange, Red, Black
}
\keyword{datasets}
